Skip complex function tests on darwin/intel.
On Mac OS X 10.7, numpy.test() fails with messages
similar to those reported on Windows in trac 1574
( http://projects.scipy.org/numpy/ticket/1574 ).

It seems the platform-provided C99 functions are also
flakey on OS X. According to the logic in trac 1574, these
tests should be skipped since the bug is in the platform
library, not numpy.

This patch skips the tests on darwin/intel, and
makes numpy.test() pass successfully on OS X 10.7.